Start project module
Create a project module
directory and a
project module js.

The to list is envisaged
to have a projects array
which will contain each
project object. Each
project object would have
a name, color and tasks
property. The tasks
property would be an
array of tasks objects.
Multiple tasks objects can
be added to the tasks array
of a project.

The project module first
initializes an empty projects
array and exports it.
The createProject factory
function takes 3 parameters.
When it returns the object,
it used the spread operator
to create a shallow copy of
the elements of
initialTasks to a new array
and assigns the array to
tasks.

That way the function can be
called with an initialTasks
array and subsequent attempts
to modify to add more tasks
to the tasks array will not
alter the original initialTasks
array.

The pushProjects function takes
in a projectObject parameter
and pushes it into the
project array.

Create a new project and push
it into the project array.