Start the tasks module
Create a tasks module directory
and a tasks module js file.

The js file first imports the
project array so that it can
push tasks into the the required
project object in the array.

Create and export a createTask
function that returns a task
object. Initialize a task object.
Create and export an addTask
function that takes a task object
and pushes it into the tasks
property of a projectObject.
(Remember that the tasks
properties are arrays.)

Push the just created task into
the tasks array of the first
element in the project array.

Log the projects array to the
console.

Rename the projects module
directory.