Refactor createTask() to Task
Convert the createTask() to
a Task class so that methods
to update the values of the
object properties can be
added to the Task class.
Exporting the Task class also
exports these methods. It
is better than exporting
individual functions.

Add the status parameter to
the constructor so that the
task can be set to completed
or pending. Add the addTask()
as a method in the Task class.
Add other methods will be
used to edit the values of
the properties of the
objects created with the
Task class.

Update task1 to include a
value for the status
property and call
the addTask method to
push task1 into
the tasks array of the first
element in the project array.