Add Task delete method
Change the name of the
addTask() method to
add() and update its
instance on line 55.

Add a delete() method.
The delete method takes
in an projectObject
parameter.
It looks for the index of
the said task in the tasks
array of the project object.

If the index exists, it
splices it out of the
tasks array of the project
object.