Render projects
First add 2 container divs
in index.html. One as a
container for
projects and the other for
tasks.

Select these divs in the
index.js

Create a function that
renders all the Projects.
The function loops through
the projects array, and
creates a projectElement div
for each element in the array.
Give each project element
a text content of the project
title and the length of its
tasks array. Give each project
element a class. Give each
project a data attribute that
corresponds to its index in
the projects array.

Append each projectElement
to the project container.

Call the function that was
just created.