Render tasks
Create a function that
takes a project object
as its parameter. It
clears the textContent of
the task container so that
anytime it is called, a new
set of tasks replace the
previous ones of the
tasks container.

Its loops through the tasks
array of the project object
and creates an edit button,
a delete button and a
task element div for each item
in the array. Set data attributes
on the edit and delete buttons
that corresponds to the index
of the corresponding task item
in the tasks array. This is
so that when these buttons
are clicked, these data
attributes will point to the
corresponding task.
Then use the for...in loop to
loop through all enumerable
properties of each task.
The hasOwn() checks if each
property is directly on the
task object and not inherited
from its prototype chain.
If true, it creates a task
element property div, adds
an appropriate class. Set the
textContent as the property
value and append the property
to the task element.

This creates a div for every task
and a div for for each
property in each task and
appends each property to
its respective task.

Append task element to the
task container.

Call the newly created function
on a project object.