Handle project clicks
Create a function that displays
a project's function onClick.

The function takes in an event
parameter and
clears the textContent of
the task container. Then it takes
the index data attribute of the
event's target(the project div that
was clicked) and saves it in a
selectedIndex variable.
It then takes this index attribute,
uses it with the projectsArray
to determine the
project that was clicked and saves
the project in a selectedProject
variable.

It returns if a project element
was not clicked because there may
be empty space and other elements
in the project container.

It then calls the renderTask function
on the selectedProject.

Add a click event listener of the
projectsContainer and set this
newly created function as a callback.