Get all edit task buttons
First update the class names
of the edit task and delete
task buttons.

Edit the handle project
clicks function such that
it takes in a callback function
as one of its parameters.
Call the callback function
after the tasks have been
rendered so that it will be
possible get a node list of
all edit task button currently
on the page.

Create an edit tasks function
that simply gets a node list of
all edit task buttons and logs
to the console.

Edit the event listener attached
to the project container, such
that it takes an anonymous
function as its second argument
and calls the handle project
clicks function. This handle
project function then calls
its event argument and another
anonymous function that calls
the edit tasks function as
its callback.

Choose to use an anonymous
function and not directly
passing a reference of editTasks
for handle project clicks because
there will be more functions
after editTasks.