Add the tasks modal
First add a load event listener to
window and call editTask when the
window loads. This is because there
is a set of default tasks that are
rendered when the page load even
before any clicks are made to any of
the projects. Thus editTask is able
to get a nodelist of all edit task
buttons on the pages once the page
loads

Add a task dialog to index.html.
This dialog will be used to add
new tasks and edit tasks. Add
a form in the dialog with a text
textarea, date picker, select, and
radio buttons inputs. Also add
2 buttons, for confirm and cancel.

In index.js attach a click event
listener to each button that shows
the task modal.