Capture the index of project
In index.js renderTasks(),
capture the index of the
project (in the projects array)
whose tasks are being
rendered.
It will be used to
pinpoint the particular
task to be edited or
deleted.

Add a data attribute which
corresponds to the index
of the project to the
edit and delete buttons

Update the data attribute
of the edit
buttons from index to
task so that it corresponds
with the project attribute.

Remove the hard coded tasks
from the project on line 7
and instead create tasks
that are instances of Task
and push them into projects.
Hard coded tasks are not
instances of the Task class
and thus the methods of
Task cannot be called on them.

Update project1 in project
module js for this reason.

Update task1 in tasks module
js with meaningful details.

In the index.html, fix little
things there and change the
status input from radio buttons
to select inputs.