Populate modal form
Update the class name of each
project element on line
27. The projects updated
from the last commits had
spaces in their titles
and classes with spaces
throw an Uncaught
InvalidCharacterError.

Move the constants concerning
the modal from inside editTasks
to the global scope because they
will be used by a createTask
function.

Add constants selecting the
modal form fields.

In the click event listener
attached to each editTaskButton
button, add the event parameter
so that its target (the button
that was clicked) can be
retrieved.

Get the projectIndex and
taskIndex attributes from the
button that was clicked set
them on the confirm button.
This is so that when the
task dialog close event is
called, it can retrieve these
values from the confirm button
and update the necessary task.

Get the selectedTask and
populate the form fields
in the modal with the
respective properties of
the selectedTask.

Thus when the modal opens,
the form fields are
populated.