Handle confirm button click event
First set the confirm button's
value on line 112.

Add an event listener on button
that preventDefault (prevents
the button from submitting the
form) and calls the close event
on the task dialog with the
confirm buttons's value as
the argument.

Update the task dialog's
close event such that the task
properties update occurs if the
task dialog's return value is
the same value as what was set
as the value of the confirm
button.

If the confirm button is
clicked, it preventDefault
and passes in its value,
as the task dialog's
returnValue - close(returnValue).
As the close event is fired,
the tasks dialog checks whether
its return value and calls the
update methods as needed.

This is just a way of closing
the dialog via the confirm
button and updating the tasks
properties in the process.