Update the display
Update the display after
the tasks properties have
been updated to reflect the
new values by calling
the renderTasks on the
project index.

Call the editTasks recursively
so that after renderTasks is
called, editTasks fetches
the nodelist of edit task buttons
and sets the click event listeners
on the these buttons and gets ready
to update the tasks properties
again.