Prepare for creating new tasks
In the index.html, add new task
button. Add a container of another
set of confirm
and cancel buttons to handle new tasks
and give their container a new tasks
buttons class. Give the container of
the other set
of buttons (now exclusively for
editing tasks) a edit tasks buttons
class. This is to be able to target
each one and hide/show as needed.

Add a projects dropdown div after the
status dropdown div. This contains
the select element for the projects.
The options of this select element
will be dynamically generated.

In the index.js, select these the
button containers. Make the confirm
and cancel buttons constants local to
the editTasks() because creating new
tasks will have its own confirm and
cancel buttons and these do not have
to be global.