Create handleCreateTasks
First, in editTasks, set the
display of the container of the
new task buttons
container to none and the
that of the edit tasks buttons
to an empty string. This is so
that when the task dialog
opens to edit task, it is
only the edit tasks confirm
and cancel buttons that
will be visible.

This function will handle
a part of creating tasks.
Create the relevant
constants before the function.

In the function, set a value
for the confirm button. This value
will be used in the dialog
close event later.

Hide the confirm and cancel buttons
for edit tasks and show the
ones for new tasks. Show the
modal and clear the form in the
modal from edit previous values
(edit tasks)

Create a default option,
and assign necessary
attributes to
it. Append it to projects field.
Create it in JS because if it
is created in the HTML, other
options appended from the JS tend
to overwrite it?

Loop through the projects array
and create a select option
for each project. Assign each,
a text content of its
corresponding title and a value
of its corresponding index.
Append each select option to
the projects field

Attach a click event listener
to new task button with this
handleCreateTasks as a callback.