Setup event listeners for new task
Define a function that sets up
event listeners for the taskDialog
close event and the confirm
and cancel buttons click events.
After defining the function,
it is called only once after
the script is loaded to register
the event listeners.

If the event listeners are set up
inside the handleCreateTasks
function, each time the function
is called (by clicking on the
new task button), a new event
listener is added without
removing any previous
listeners for the same event.
In this case, two event
listeners will be registered
for the confirm button when
the new task button is
clicked the second time, and
a task would be added twice.

In renderProjects(), first set
the project container so that
every time it is called, a new
list of projects replace
the previous ones.

In the setup event listeners
function, set up the taskDialog
close event such that it first
checks its return value for a
string and if true creates a
new task with the form field
values and adds the new task
at the beginning of the
tasks array of the project
selected in the projects
field. It then renders projects,
renders the tasks of the said
project and calls the editTasks().

Call the setup event listeners
function.