Setup event listeners for buttons
In the event listener function,
setup a click event listener for
the confirm button so that it
calls close() and passed a return
value for the taskDialog - just
as in editTasks.

Setup a click event listener for
the cancel button so that it first
sets a different return value for
the tasksDialog before closing it.
If a different returnValue is
not set and it calls the tasksDialog
to close(), the dialog uses the
last set returnValue (which
is "thisIsForNewTasks" from
the previous submit). Thus the
cancel button would create a
new task after a previous
submit. Setting a different
returnValue for the tasksDialog
prevents this behavior.