Edit editTasks
Rename the names of the
constants of confirm and cancel
buttons and update all their
instances to reflect the new name.
Move them to a global scope because
an event listener function
will need them later.

Change the returnValue of tasksDialog
in the cancel button click event
listener to prevent the behavior
described in the previous commit.