Add update project
In the Task class, define
a method that updates
the project of a task.

The method takes in the
previous object and the
new object as parameters.
The method than gets the
index of the particular
task from the tasks
array of the previous
object. It then takes the
index, plugs it into the
tasks array of the previous
object and adds it as the
first element in the
tasks array of the new
object. It then calls the
delete method with the
previous project object
as an argument.