Define create select options function
Hoist the definition of projectsField
higher on 118 so that a function can
call it.

Create the function and move the
logic that creates select options
from handle create tasks to it.
Have the function return an
object containing the
select options.

Create this function because it
will be called by editTasks()
so that a user can update a
task's project while updating the
task.

Call this function in the
handle create tasks function.