Implement updateProject()
First in renderTasks(),
editTasks(), and
setupEventListenersForEditTasks()
update
projectIndex definitions and all
its references to oldProjectIndex.
This is because a new constant
that would hold and reference
index of new Project where the task
would be updated to would be
created.

In editTasks() set the value of
the projectsField to the
old project index of the task.
This is so that when the dialog
opens, the project field would
display the tasks's project.
Add a change event listener to
the project field such upon
change, it sets a data attribute
to the confirm button. The value
of the data attribute is the new
value od the project field (the
new project that was selected)

In the setupEventListenersForEditTasks(),
it the dialog event, it retrieves the
value of the attribute that was set on
the confirm button and creates a
new project index constant from it.

Thus during updating a task, it checks
if new project index exists and if true,
it calls the updateProject method with
the relevant parameters and renders
the tasks of the new project.
Else (everything may have edited but the
task's project was not) it renders the
tasks of the old/original project.
Remove the data attribute that was set
on the confirm button. If the data
attribute was not removed,
the submit button
would hold the last new project
index that was set on it.
Thus next time the dialog is
opened for tasks of another project,
clicking on the confirm button would
still call the updateProject method
and update the tasks to the project
with the index of this new project
index value that has been set on it.
This happens if there is not 'change'
in the project field.