Implement delete tasks
Add a deleteModal in index.html.
This modal will be used to confirm
or cancel deleting tasks and
projects. Add two sets of buttons
in the modal, confirm and cancel,
for tasks and projects.

In index.js, implement delete tasks
edit tasks as implemented.
Define and call a delete tasks
function and an
appropriate event listener function
to handle the deleteDialog and
its submit button and cancel buttons.

Call the delete tasks function
in the event listener function
before the deleteDialog is closed
so that after deleting a task,
event listeners are set on the
delete buttons over again.