Call delete tasks function
Call the delete tasks function
in the event listener on
project container (97) so that
when a project in the
container is clicked and
its tasks are rendered, the
delete tasks sets an event
listener on all delete buttons.

Also call it in the event
listener functions for
edit tasks (220) and
new tasks (305)
for this same purpose.