Set event listener on project divs
With the click event listener
set on the project container,
if a click happens on any element
that is not
a project element(inside the
project container),
it returns (line 107)
and does not
call the renderTasks function.
This makes the previous tasks
rendered to hide.

From line 114, Select the projects
themselves
(target them by class) and
add the event listener on
the node list instead, with
a for loop.

Now clicking on the other
elements like the edit
and delete project buttons
in the project container
do not hide the previous
tasks.