Put event listeners in a function
As the event listeners are set on
the node list, after any action
(edit, delete, etc.) happens,
clicking on a project element
does not render its tasks.

Put the event listeners in a
function and call the function
so that the event listeners
are set on page load.

Call the function inside
call helper functions, so that
this after any action, the
function is called and the
event listeners on
project divs are set again.