Add create project feature
In index.js,
set up handleCreateProject()
and
setupEventListenersForCreateProjects()
just was done for new tasks. On
line 543, after the new project has
been created, render its task
(render the tasks of the first
element in the projects array
because the newProject was
'unshifted' into the projects array).

In index.html, add the new project
button and update the ids of the
project dialog buttons to
correspond with the ids used to
target them in the index.js