Add delete project feature
In index.js,
set up deleteProjects()
and
setupEventListenersForDeleteProjects()
just was done for delete tasks.

In the deleteDialog close event,
before the delete() method is
called on the selectedObject,
on line 603, check if the
the index of the object
to be deleted from projectArray
is the last
(the last element). If so,
render the tasks of the
first element, else render the
tasks of the next element
(first convert the projectIndex
to a number). The else block
makes it that if project
element with index 1 is deleted,
the tasks of project element with
index 2 is rendered.

Add deleteProjects() in the
list of functions inside
callHelperFunctions() so
that the event listeners
set on the delete project
buttons will be reset after
after any action (edit,
delete) on tasks and projects.

In index.html, update the id
of the delete projects cancel
button to
correspond with the id used to
target it in the index.js