Handle empty project array
When the last project is
deleted from the projectArray,
its tasks are still rendered
on the screen.

Define a isProjectArrayEmpty()
function returns
true or false depending on
whether the project array
is empty or not. Define a
clearTasksContainer() that
checks isProjectArrayEmpty()
clears the
task container accordingly.
Define handleNewTaskButton()
that disables
the add new task button if
the isProjectArrayEmpty().

Call the last 2 aforementioned
functions in
callHelperFunctions().

In renderTasks() output a
message to the console
if the said project
does not have any tasks.
Later, add a user friendly
display.