Define hide/show div function
Define a function on line 39,
that has has 2 parameters.
it hides the display of one
and show the display of
the second.

Select the divs for the deleteTasks
and deleteProjects buttons on
lines 388 and 389. Call this
function in deleteTasks
and deleteProjects to hide
and show the necessary buttons
since both of them use the
deleteDialog and both of
their buttons are on the
dialog.