Call hide/show div function
Use the hide/show div function
call and the relevant arguments
to hide/show button containers
in editTask, handleCreateTasks
editProject, and
handleCreateProjects. Use the
function call to replace
the previous method to hide/show
button containers.

Update the names of
deleteTasksButtonsDiv and
deleteProjectButtonsDiv and
their references to be
consistent with the naming
convention of
other button containers.