Make deadline a Date object
In the Task class constructor,
make the deadline property
a new Date object and have the
updateDeadline() method update
with a new Date property.

This is to be able to sort the
tasks according to their
lines and perform other
formatting on the deadline date.

The full date and time now
displays in the tasks container
but when the dialog is opened
for edit tasks, the deadlineField
no longer holds the date of
selectedTask because of the
new date format. Should try
to fix that later.