Sort project array by deadline
Define a sortArrayByDeadline
function that takes
in an array as an argument
and sorts its the items in
its elements'
task property (array) by their
deadline.

Move the renderTasks() call
on project to line 171, in the
part that handles what happens
on window load - (Make sure
to call it before the editTasks()
else editTasks() not work on
window load.) This makes it
to render the tasks on window
load and as the
sortArrayByDeadline is called
on page 673, the tasks are
sorted by their deadlines
on window load.

Call the sortArrayByDeadline
inside callHelperFunctions.