Update and retrieve from localStorage
On line 3, use an import specifier
(import alias) to import the
projectsArray. The variable which
will hold the instance of the
function that will check the
localStorage and return the
localStorage array or the
initialArray
will be named projectsArray.
Rename the instances on lines
17, 18, 19 and 21 to
the new alias

This is so that the app can work with
this array and avoid renaming all
instances of the projectArray to
a new name.

On line 58, define a
updateLocalStorage() that saves
the initialArray to the
localStorage. On line 62,
define a
fetchLocalStorageArrayOrDefault().
This function defines a constant
that tries to retrieve the array
that was saved by
updateLocalStorage()
from localStorage or an empty
array. The empty array is to
prevent it from crashing if
there is no array in localStorage.
In the function, define a variable
that is the result of a check if the
array in local storage is empty or not
(If empty, initial array, else
local storage array).
Return the variable.

On line 21, assign the result of
fetchLocalStorageArrayOrDefault()
to a constant (that is named
projectsArray for reasons stated
above).