Edit fetchLocalStorageArrayOrDefault()
Now, the elements of the
local storage array have
been stripped of their
class methods when they
are retrieved from the
localStorage. This is due
to the behavior of
JSON.stringify() and
JSON.parse(). Project
and Task methods cannot
be called on any of them.

Use the map() method to
loop through each element
in the local storage array
and
construct a new Project class
instance of each of them.
When it gets to the 3rd
argument in the Project
constructor (initialTasks),
it also loops through its
elements and creates a
new Task class instance
of each of them. The map()
returns new arrays with
the elements as Projects
and Task
class instances respectively.

Now the Projects and Task
methods can be called on
the elements of the
local storage arrays.