[Update storage after new project
The local storage was not being
updated after a project was
added despite updateLocalStorage()
being called in callHelperFunctions()
and helper functions called
after a new project is added.

Call the updateLocalStorage()
after the new project is added
(unshifted)
to the project array, before
the new project is rendered.
This updates the local storage
after a new project is added.