Set validity rules for edit tasks
Define a checkInputValidity
function that takes in
any number of args, loop through
those args, trim the whitespace
and check if the value is empty.
If so, its sets a custom validity
message and reports the validity
message. Else it sets an empty
string as the validity message.

Call this function inside the
editTasksConfirmButton click
event listener such that it first
checks the validity of the
required input fields and if
they are all valid, it calls
the close() method on the
dialog.