Set validity rules for new tasks
Call the checkInputValidity() for
newTasksConfirmButton click event
just as it was called in the
editTasksConfirmButton click
event. Did not add the projectsField
to the args because
checkInputValidity() was not
working for it, maybe because
the projectsField was dynamically
generated.

Since the checkInputValidity()
was not working for the
projectsField, remove the
Select project default option
feature from createSelectOptions()
so that the projects field will
default to the first option
in the field.