Modify checkInputValidity()
For the projectsField in new tasks,
it seems as some browsers like
chrome does not update the value
property of a select element when
a disabled and hidden option is
selected. They instead keep track of
the selected option internally.
This seems to be the reason why the
projectsField was passing the test
in the if block of
checkInputValidity() even if nothing
is selected. If another option
is not selected, browser just selects
it, even though it is disabled
and hidden.

Modify checkInputValidity() such
that, it takes in one more
arg - a special select field.
It now searches for and returns
an option element
that is selected within this field.

In the for loop, it now checks
if other inputs are empty of
if this option that is selected
is disabled. Thus, if the user
does not select and option and
leaves it on the default Select
Project then it sets the
validity message and reports it.

Implement it in editTasks, newTasks,
editProjects and newProjects
with projectsField and
projectColorField
being the specialSelectField
for tasks and projects
respectively.