Modify updateLocalStorage()
Local storage was updating
as expected upon create,
update and delete of
tasks and projects but
upon refresh it defaults
to the initial projects array.

Modify updateLocalStorage()
such that localStorage
is updated based on the
live projectsArray and not
the initialProjectsArray.
Now localStorage is not
overwritten with the initial
array on page reload.

Performing any operations
is working directly with
projectsArray which may
be the array from
localStorage or
initialProjectsArray (if
localStorage array is empty)
So this is basically saying,
take whatever is in
projectsArray, save it in
localStorage so that
fetchLocalStorageArrayOrDefault()
can return it.