Fix renderTasks() on load
On load, the tasks of project2
were being rendered. When I edit
a project2 task, reload the page
and try to edit another
project2 task, it
throws an error, because
oldProjectIndex
(data-old-project) is now -1. If
I click on another project
to render its tasks at this
point before I click the edit
button, it works data-old-project
is now 1 and it works as
expected.

Have it render the tasks of
the first project in the
projectsArray on window load.
Now it does not render a
particular project, it is dynamic
and renders whatever is the
first project in projectsArray.
This fixes the issue.